+++ src/types.rs
@@ -8,7 +8,7 @@ pub enum Error<E> {
 }
 
 /// RGB converter gain
-#[derive(Debug, Clone, Copy, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum RgbCGain {
     /// 1x gain
     _1x,
@@ -23,7 +23,7 @@ pub enum RgbCGain {
 /// RGB converter interrupt persistence
 ///
 /// This controls the RGB converter interrupt generation rate.
-#[derive(Debug, Clone, Copy, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum RgbCInterruptPersistence {
     /// Every RGBC cycle generates an interrupt.
     Every,
@@ -60,7 +60,7 @@ pub enum RgbCInterruptPersistence {
 }
 
 /// Result of measurement of all channels
-#[derive(Debug, Clone, Copy, PartialEq)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct AllChannelMeasurement {
     /// Red channel measurement.
     pub red: u16,